Refactor for improved method backwards compatibility
